tests/wpt/web-platform-tests/feature-policy/experimental-features/intrinsicsize-with-unsized-media.tentative.https.sub.html

<!DOCTYPE html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

const srcs = [
  "/feature-policy/experimental-features/resources/image.svg",
  "/feature-policy/experimental-features/resources/image.jpg",
  "/feature-policy/experimental-features/resources/image.png",
];

// Set new attribute, and wait til the media element is repainted.
function updateAttribute(e, attribute, value) {
  return new Promise(resolve => {
    requestAnimationFrame(() => requestAnimationFrame(() => resolve()));
    e.setAttribute(attribute, value);
  });
}

// Test intrinsicSize attribute with image element.
for (var src of srcs) {
  promise_test(async() => {
    var img = document.createElement('IMG');
    document.body.appendChild(img);
    img.src = src;
    await new Promise(resolve =>
      img.addEventListener('load', () => resolve(), {once: true}));
    assert_equals(img.width, 300, 'width');
    assert_equals(img.height, 150, 'height');
    assert_equals(img.naturalWidth, 300, 'naturalWidth');
    assert_equals(img.naturalHeight, 150, 'naturalHeight');

    await updateAttribute(img, 'width', '800');
    assert_equals(img.width, 800, 'width');
    assert_equals(img.height, 400, 'height');
    assert_equals(img.naturalWidth, 300, 'naturalWidth');
    assert_equals(img.naturalHeight, 150, 'naturalHeight');

    await updateAttribute(img, 'intrinsicSize', '400 x 500');
    assert_equals(img.width, 800, 'width');
    assert_equals(img.height, 1000, 'height');
    assert_equals(img.naturalWidth, 400, 'naturalWidth');
    assert_equals(img.naturalHeight, 500, 'naturalHeight');

    await updateAttribute(img, 'style', 'height:800px;');
    assert_equals(img.width, 800, 'width');
    assert_equals(img.height, 800, 'height');
    assert_equals(img.naturalWidth, 400, 'naturalWidth');
    assert_equals(img.naturalHeight, 500, 'naturalHeight');

    await updateAttribute(img, 'width', '');
    assert_equals(img.width, 640, 'width');
    assert_equals(img.height, 800, 'height');
    assert_equals(img.naturalWidth, 400, 'naturalWidth');
    assert_equals(img.naturalHeight, 500, 'naturalHeight');

    await updateAttribute(img, 'intrinsicSize', '');
    assert_equals(img.width, 1600, 'width');
    assert_equals(img.height, 800, 'height');
    assert_equals(img.naturalWidth, 300, 'naturalWidth');
    assert_equals(img.naturalHeight, 150, 'naturalHeight');

    await updateAttribute(img, 'style', '');
    assert_equals(img.width, 300, 'width');
    assert_equals(img.height, 150, 'height');
    assert_equals(img.naturalWidth, 300, 'naturalWidth');
    assert_equals(img.naturalHeight, 150, 'naturalHeight');
  }, 'Test image with src=' + src);
}

// Test intrinsicSize attribute with video element.
promise_test(async() => {
  var video = document.createElement('video');
  document.body.appendChild(video);
  video.src = "/feature-policy/experimental-features/resources/video.ogv";
  await new Promise(resolve =>
  video.addEventListener('canplaythrough', () => resolve(), {once: true}));
  assert_equals(video.getBoundingClientRect().width, 300, 'width');
  assert_equals(video.getBoundingClientRect().height, 150, 'height');
  assert_equals(video.videoWidth, 300, 'videoWidth');
  assert_equals(video.videoHeight, 150, 'videoHeight');

  await updateAttribute(video, 'width', '800');
  assert_equals(video.getBoundingClientRect().width, 800, 'width');
  assert_equals(video.getBoundingClientRect().height, 400, 'height');
  assert_equals(video.videoWidth, 300, 'videoWidth');
  assert_equals(video.videoHeight, 150, 'videoHeight');

  await updateAttribute(video, 'intrinsicSize', '400 x 500');
  assert_equals(video.getBoundingClientRect().width, 800, 'width');
  assert_equals(video.getBoundingClientRect().height, 1000, 'width');
  assert_equals(video.videoWidth, 400, 'videoWidth');
  assert_equals(video.videoHeight, 500, 'videoHeight');

  await updateAttribute(video, 'style', 'height:800px;');
  assert_equals(video.getBoundingClientRect().width, 800, 'width');
  assert_equals(video.getBoundingClientRect().height, 800, 'height');
  assert_equals(video.videoWidth, 400, 'videoWidth');
  assert_equals(video.videoHeight, 500, 'videoHeight');

  await updateAttribute(video, 'width', '');
  assert_equals(video.getBoundingClientRect().width, 640, 'width');
  assert_equals(video.getBoundingClientRect().height, 800, 'height');
  assert_equals(video.videoWidth, 400, 'videoWidth');
  assert_equals(video.videoHeight, 500, 'videoHeight');

  await updateAttribute(video, 'intrinsicSize', '');
  assert_equals(video.getBoundingClientRect().width, 1600, 'width');
  assert_equals(video.getBoundingClientRect().height, 800, 'height');
  assert_equals(video.videoWidth, 300, 'videoWidth');
  assert_equals(video.videoHeight, 150, 'videoHeight');

  await updateAttribute(video, 'style', '');
  assert_equals(video.getBoundingClientRect().width, 300, 'width');
  assert_equals(video.getBoundingClientRect().height, 150, 'height');
  assert_equals(video.videoWidth, 300, 'videoWidth');
  assert_equals(video.videoHeight, 150, 'videoHeight');
}, 'Test video');
</script>
</body>
